Refactor CPIO handlers to the new class-based implementation
The module structure have been simplified by moving them to the same module,
and there is a new mixin class for the new ASCII based formats, which
contains the elements which are the same for both.
