--- unblob/handlers/archive/cpio/__init__.py
@@ -1,4 +1,0 @@
-from .cpio_default import BinaryHandler
-from .cpio_portable_ascii import PortableASCIIHandler
-from .cpio_portable_ascii_crc import PortableASCIIWithCRCHandler
-from .cpio_portable_old_ascii import PortableOldASCIIHandler
